requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.10.0
pydantic-settings>=2.1.0

# Azure AI Services
azure-ai-formrecognizer==3.3.2
azure-search-documents==11.4.0
azure-identity==1.15.0
azure-core==1.29.6
openai>=1.12.0

# LangChain
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.17
langchain-text-splitters>=0.0.1

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
chardet==5.2.0
python-multipart==0.0.6

# Vector Store & Embeddings
faiss-cpu>=1.12.0
tiktoken>=0.5.2

# Web Framework
jinja2==3.1.3
aiofiles==23.2.1

# Utilities
numpy>=1.26.0
pandas>=2.2.0
requests>=2.31.0
pytz>=2023.3
